+++ tests/full/test_full.tf
@@ -15,8 +15,8 @@ module "main" {
   source = "../.."
 
   name          = "LEAF105"
-  id            = 105
-  pod           = 2
+  node_id       = 105
+  pod_id        = 2
   serial_number = "ABCDEFGHIJKLMN"
   role          = "leaf"
   type          = "tier-2-leaf"